Load & Save Account with passwords

+++ Form1.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -68,6 +69,9 @@ namespace password_manager
 
         private void removePassword_Click(object sender, EventArgs e)
         {
+            if(this.passwordGridList.CurrentCell == null)
+                return;
+
             int selectedRow = this.passwordGridList.CurrentCell.RowIndex;
 
             PasswordRepository.RemovePassword(passwordList[selectedRow]);
@@ -78,9 +82,6 @@ namespace password_manager
         private void OnNewPasswordCreated(object sender, EventArgs e)
         {
             LoadPasswords();
-
-            Console.WriteLine("New password created");
-            Console.WriteLine("Password list count: " + passwordList.Count);
         }
 
         private string hashPassword(string password)
@@ -107,7 +108,7 @@ namespace password_manager
 
             // toggle password visibility
             string toggleVisible = isPasswordVisible[selectedRow] ? hashPassword(passwordList[selectedRow].getPassword()) : passwordList[selectedRow].getPassword();
-            passwordGridList[1, selectedRow].Value = toggleVisible; // [1, selectedRow] is the same as [selectedRow, 1
+            passwordGridList[1, selectedRow].Value = toggleVisible;
 
             // update visibility array
             isPasswordVisible[selectedRow] = !isPasswordVisible[selectedRow];
@@ -115,14 +116,12 @@ namespace password_manager
 
         private void toggleAllPasswords(object sender, EventArgs e)
         {
-
-
             int rowCount = passwordGridList.Rows.Count;
 
             if(rowCount <= 0)
                  return;
 
-            for(int i = 0; i < rowCount; i++)
+            for(int i = 0; i < rowCount; ++i)
             {
                 string passwordType = isAllPasswordsVisible ? hashPassword(passwordList[i].getPassword()) : passwordList[i].getPassword();
 
@@ -135,7 +134,57 @@ namespace password_manager
 
         private void loadAccount_Click(object sender, EventArgs e)
         {
+            openAccountDialog.ShowDialog();
 
+            string filePath = openAccountDialog.FileName;
+
+            if(filePath == null) return;
+            if(filePath == "") return;
+
+            string fileName = System.IO.Path.GetFileName(filePath);
+
+            loadedAccount.Text = fileName;
+
+            string[] lines = System.IO.File.ReadAllLines(filePath);
+
+            foreach(string line in lines)
+            {
+                string[] account = line.Split(',');
+
+                Password newAccount = new Password(account[1], account[0], account[2]);
+
+                PasswordRepository.AddPassword(newAccount);
+            }
+
+            LoadPasswords();
+        }
+
+        private void saveAccount_Click(object sender, EventArgs e)
+        {
+
+            var ifFolderExist = Directory.Exists("accounts");
+                
+            if(!ifFolderExist)
+            {
+                System.IO.Directory.CreateDirectory("accounts");
+            }
+
+            //create new file inside accounts path
+            string filePath = "accounts/" + DateTime.Now.ToString("yyyy-MM-dd") + ".txt";
+            if(File.Exists(filePath))
+            {
+                File.Delete(filePath);
+            }
+
+            using (StreamWriter sw = File.CreateText(filePath))
+            {
+                foreach(Password password in passwordList)
+                {
+                    sw.WriteLine(password.getUsername() + "," + password.getPassword() + "," + password.getPlatform());
+                }
+            }
+
+            MessageBox.Show("Account saved successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information, MessageBoxDefaultButton.Button1);
         }
     }
 }
